Move Spanners-related types to spanners file
