+++ src/rendering/spanners.ts
@@ -1,7 +1,7 @@
 import { Beam, BeamRendering } from './beam';
 import { Slur, SlurRendering } from './slur';
 import { Tuplet, TupletRendering } from './tuplet';
-import { BeamFragment, SlurFragment, SpannerFragment, TupletFragment } from './types';
+import * as vexflow from 'vexflow';
 
 /** The result of rendering spanners. */
 export type SpannersRendering = {
@@ -9,6 +9,59 @@ export type SpannersRendering = {
   beams: BeamRendering[];
   tuplets: TupletRendering[];
   slurs: SlurRendering[];
+};
+
+/** Describes what part of the spanner lifecycle a spanner fragment is in. */
+export type SpannerFragmentPhase = 'unspecified' | 'start' | 'continue' | 'stop';
+
+/**
+ * Represents a piece of a spanner.
+ *
+ * Spanners are structures that involve groups of notes.
+ *
+ * Examples:
+ *   - beams
+ *   - tuplets
+ *   - slurs
+ */
+export type SpannerFragment = BeamFragment | TupletFragment | SlurFragment;
+
+/** Represents a piece of a beam. */
+export type BeamFragment = {
+  type: 'beam';
+  phase: SpannerFragmentPhase;
+  vexflow: {
+    stemmableNote: vexflow.StemmableNote;
+  };
+};
+
+/** Represents a piece of a tuplet. */
+export type TupletFragment =
+  | {
+      type: 'tuplet';
+      phase: 'start';
+      vexflow: {
+        note: vexflow.Note;
+        location: vexflow.TupletLocation;
+      };
+    }
+  | {
+      type: 'tuplet';
+      phase: 'unspecified' | 'stop';
+      vexflow: {
+        note: vexflow.Note;
+      };
+    };
+
+/** Represents a piece of a slur. */
+export type SlurFragment = {
+  type: 'slur';
+  phase: SpannerFragmentPhase;
+  slurNumber: number;
+  vexflow: {
+    note: vexflow.Note;
+    keyIndex: number;
+  };
 };
 
 /**
